PER, PBRチャートを表示
