+++ app/javascript/packs/application.js
@@ -5,7 +5,12 @@ import React from 'react';
 
 document.addEventListener("DOMContentLoaded", () => {
   ReactDOM.render(
-    <App code={4368} />,
+    <App code={4368} indices={"PER"} />,
     document.querySelector("#per-chart")
+  );
+
+  ReactDOM.render(
+    <App code={4368} indices={"PBR"} />,
+    document.querySelector("#pbr-chart")
   );
 });
