+++ app/javascript/packs/components/App.jsx
@@ -1,78 +1,40 @@
 import React from "react";
 import ChartContainer from "./ChartContainer";
+import axios from "axios";
+import { useEffect, useState } from "react";
+import { currentConfigFn, entireConfigFn } from "./ChartConfig";
 
-const App = ({ code }) => {
+const App = ({ code, indices }) => {
+  const [currentConfig, setCurrentConfig] = useState();
+  const [entireConfig, setEntireConfig] = useState();
+
+  useEffect(() => {
+    getData(code, indices).then(data => {
+      const currentData = data.current_year;
+      let labels = currentData.x_label || []
+      let points = (currentData.data || []).map((point) => { return {x: point[0], y: point[1]} });
+      setCurrentConfig(currentConfigFn(labels, points));
+
+      const entireData = data.entire_period;
+      labels = entireData.x_label || []
+      points = (entireData.data || []).map((point) => { return {x: point[0], y: point[1]} });
+      setEntireConfig(entireConfigFn(labels, points));
+    });
+  }, []);
+
   return (
     <div>
-      <ChartContainer code={code} config={currentConfig} >年初来</ChartContainer>
+      <ChartContainer indices={indices} config={currentConfig} >年初来</ChartContainer>
       <div className="clearfix" />
-      <ChartContainer code={code} config={entireConfig} >全期間</ChartContainer>
+      <ChartContainer indices={indices} config={entireConfig} >全期間</ChartContainer>
     </div>
   );
 };
 
-const currentConfig = (labels, points) => {
-  return {
-    data: {
-      datasets: [{
-        borderWidth: 2,
-        pointHitRadius: 7,
-      }],
-    },
-    options: {
-      scales: {
-        xAxes: [{
-          type: 'linear',
-          ticks: {
-            callback: (value, i, values) => labels[value],
-            max: labels.length,
-            stepSize: 60
-          }
-        }],
-      },
-      tooltips: {
-        callbacks: {
-          label: (item, data) => {
-            const point = points[item.index]
-            const value = point.y
-            return ` ${value} 倍（${labels[point.x]}）`
-          },
-        },
-      },
-    }
-  };
-}
-
-function entireConfig(labels, points) {
-  return {
-    labels: ['2011', '2012', '2013', '2014', '2015', '2016', '2017', '2018', '2019', '2020'], // TODO
-    data: {
-      datasets: [{
-        borderWidth: 1.5,
-        pointHitRadius: 2,
-      }],
-    },
-    options: {
-      scales: {
-        xAxes: [{
-          type: 'time',
-          time: {
-            unit: 'year',
-            stepSize: 1,
-          },
-        }],
-      },
-      tooltips: {
-        callbacks: {
-          label: (item, data) => {
-            const point = points[item.index]
-            const value = point.y
-            return ` ${value} 倍（${point.x}）`
-          },
-        },
-      },
-    }
-  };
+async function getData(code, indices) {
+  const url = `/stock_prices/${code}/${indices.toLowerCase()}.json`;
+  const response = await axios.get(url);
+  return response.data;
 }
 
 export default App;
