+++ app/javascript/packs/components/ChartContainer.jsx
@@ -2,17 +2,17 @@ import React from "react";
 import Chart from "./Chart";
 import { useState } from "react";
 
-const ChartContainer = ({ code, config, children }) => {
+const ChartContainer = ({ indices, config, children }) => {
   const [max, setMax] = useState();
 
   return (
     <>
       <h5 className="chart-header">{ children }</h5>
       <div className="col-7 per-container">
-        <Chart code={code} option={config} max={max} />
+        <Chart config={config} max={max} />
       </div>
       <div className="operator-container">
-        <div>PERの最大値</div>
+        <div>{ indices }の最大値</div>
         <select onChange={(e) => { setMax(maxValue(e)) }} className="select-per form-control">
           { selectOptions([null, 10, 15, 20, 30, 50, 100]) }
         </select>
